add_op_unittest
